Extract text sections styling
