+++ engine/lib/phx/src/ui/hmgui/text/mod.rs
@@ -3,6 +3,7 @@ mod text_context;
 mod text_cursor_rect;
 mod text_data;
 mod text_render;
+mod text_section_style;
 mod text_selection;
 mod text_style;
 mod text_view;
@@ -11,6 +12,7 @@ pub use text_alignment::*;
 pub use text_context::*;
 pub use text_cursor_rect::*;
 pub use text_data::*;
+pub use text_section_style::*;
 pub use text_selection::*;
 pub use text_style::*;
 pub use text_view::*;
